I learned how to use Indirect Recursive Function

+++ daily-exercises/2025-08-28/factorial_recursionc.c
@@ -14,9 +14,9 @@ int factorial (int n)
 
 int main(void)
 {
-    unsigned long long int number;
+    int number;
     printf("Please enter a number: ");
-    scanf("%lld", &number);
-    printf("%lld", factorial(number));
+    scanf("%d", &number);
+    printf("%d", factorial(number));
     return 0;
 }
